1. 将扩展部分代码迁移到common中，便于引擎和其他扩展调用
