+++ handler/Cargo.toml
@@ -14,7 +14,3 @@ serde_derive = "1.0.207"
 serde_json = "1.0.124"
 # 异步运行时库
 tokio = { version = "1.39.2", features = ["full"] }
-# 加载dll
-libloader = "0.1.4"
-# 加载 so和dylib
-libloading = "0.8.5"
