--- handler/src/extension/dll/mod.rs
@@ -1,1 +1,0 @@
-pub mod interface;
