--- handler/src/extension/mod.rs
@@ -1,7 +1,0 @@
-mod jar;
-pub mod dll;
-mod py;
-mod so;
-pub mod interface;
-mod dylib;
-mod common;
