--- handler/src/extension/py/interface.rs
@@ -1,7 +1,0 @@
-// use std::path::Path;
-//
-// use crate::core::common::logger::interface::debug;
-//
-// pub fn load_py_extension(path: &Path) {
-//     debug(format!("Load dll extension from {}", path.display()).as_str())
-// }
