--- handler/src/extension/so/mod.rs
@@ -1,1 +1,0 @@
-pub mod interface;
