+++ handler/src/lib.rs
@@ -1,3 +1,2 @@
 pub mod handler;
 pub mod tools;
-pub mod extension;
